Common base class for struct usage
